refactor(admin): extract LoadingSkeleton to its own file
CommunityAdmin.tsx was already over the oversized-file limit; hoisting
LoadingSkeleton to module scope in the previous commit pushed it from
632 to 639 lines, which the quality gate flagged as a regression.

Extracting LoadingSkeleton to CommunityAdminLoadingSkeleton.tsx brings
CommunityAdmin.tsx to 579 lines (net -53 vs the pre-PR baseline) and
gives the skeleton component its own module-scope home.
